requirements.txt

# Core Django Framework
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.3.1
django-filter==25.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching
django-redis==6.0.0
redis==7.0.1

# Authentication and Security
descope==1.7.11
PyJWT==2.6.0

# Environment and Configuration
python-dotenv==1.0.0
python-decouple==3.8

# Payment Processing
razorpay==2.0.0

# Production Server and Static Files
gunicorn==23.0.0
whitenoise==6.11.0

# Task Scheduling
APScheduler==3.10.1

# Data Processing and Analysis
pandas==2.3.2
numpy==2.3.3
matplotlib==3.10.6
seaborn==0.13.2

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.11.7
pdfminer.six==20250506
pypdfium2==4.30.0
python-docx==1.2.0
openpyxl==3.1.5

# XML and Data Parsing
lxml==6.0.1
et_xmlfile==2.0.0

# Serialization and Validation
marshmallow==3.26.1
marshmallow-enum==1.5.1
dataclasses-json==0.5.8
email-validator==2.3.0

# HTTP and Networking
requests==2.28.2
urllib3==1.26.12
dnspython==2.7.0

# Cryptography and Security
cryptography==39.0.1
cffi==1.17.1
pycparser==2.22

# Date and Time
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
tzlocal==5.3.1

# Image Processing
pillow==11.3.0

# Plotting and Visualization
fonttools==4.60.0
contourpy==1.3.3
cycler==0.12.1
kiwisolver==1.4.9

# Utilities and Helpers
packaging==25.0
pyparsing==3.2.4
six==1.17.0
setuptools==68.2.2
wheel==0.45.1
toml==0.10.2
semantic-version==2.10.0
liccheck==0.9.2

# Type Checking and Extensions
typing_extensions==4.15.0
typing-inspect==0.9.0
mypy_extensions==1.1.0

# Core Python Libraries
asgiref==3.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
sqlparse==0.5.3